# Improve `Oj.load` performance