@ohler55 2021-08-05
Good to know. I'll take advantage of that knowledge to update other places in the new parser as well.